// src/components/molecules/foodMenu/types.ts
export interface FoodMenuProps {
  categories?: any
  menuBackground?: string
  menuCategoryContainerBorder?: string
  menuCategoryContainerBorderRadius?: string
  menuCategoryContainerBoxShadow?: string
  menuCategoryContainerDisplay?: string
  menuCategoryContainerFlexFlow?: string
  menuCategoryContainerMaxWidth?: string
  menuCategoryContainerOrderNumber?: string
  menuCategoryContainerWidth?: string
  menuCategoryDescriptionColor?: string
  menuCategoryDescriptionDisplay?: string
  menuCategoryDescriptionFontFamily?: string
  menuCategoryDescriptionFontSize?: string
  menuCategoryDescriptionFontWeight?: string
  menuCategoryDescriptionLetterSpacing?: string
  menuCategoryDescriptionLineHeight?: string
  menuCategoryDescriptionMargin?: string
  menuCategoryDescriptionMaxWidth?: string
  menuCategoryDescriptionPadding?: string
  menuCategoryDescriptionTextAlign?: string
  menuCategoryDescriptionTextTransform?: string
  menuCategoryItemsContainerDisplay?: string
  menuCategoryItemsContainerFlexFlow?: string
  menuCategoryItemsContainerJustifyContent?: string
  menuCategoryItemsContainerMargin?: string
  menuCategoryItemsContainerWidth?: string
  menuCategoryTitleColor?: string
  menuCategoryTitleDisplay?: string
  menuCategoryTitleFontFamily?: string
  menuCategoryTitleFontSize?: string
  menuCategoryTitleFontWeight?: string
  menuCategoryTitleLetterSpacing?: string
  menuCategoryTitleMargin?: string
  menuCategoryTitlePadding?: string
  menuCategoryTitleTextAlign?: string
  menuCategoryTitleTextTransform?: string
  menuDisplay?: string
  menuFlexFlow?: string
  menuItemBackground?: string
  menuItemBorder?: string
  menuItemBorderRadius?: string
  menuItemBoxShadow?: string
  menuItemDescriptionColor?: string
  menuItemDescriptionFontFamily?: string
  menuItemDescriptionFontSize?: string
  menuItemDescriptionFontWeight?: string
  menuItemDescriptionLetterSpacing?: string
  menuItemDescriptionLineHeight?: string
  menuItemDescriptionMaxWidth?: string
  menuItemDescriptionPadding?: string
  menuItemDescriptionTextAlign?: string
  menuItemDescriptionTextTransform?: string
  menuItemDisplay?: string
  menuItemFlexBasis?: string
  menuItemFlexFlow?: string
  menuItemMargin?: string
  menuItemMaxWidth?: string
  menuItemNameColor?: string
  menuItemNameFontFamily?: string
  menuItemNameFontSize?: string
  menuItemNameFontWeight?: string
  menuItemNameLetterSpacing?: string
  menuItemNameLineHeight?: string
  menuItemNameMaxWidth?: string
  menuItemNamePadding?: string
  menuItemNameTextAlign?: string
  menuItemNameTextTransform?: string
  menuItemPadding?: string
  menuItemPriceColor?: string
  menuItemPriceFontFamily?: string
  menuItemPriceFontSize?: string
  menuItemPriceFontWeight?: string
  menuItemPriceLetterSpacing?: string
  menuItemPriceLineHeight?: string
  menuItemPriceMaxWidth?: string
  menuItemPricePadding?: string
  menuItemPriceTextAlign?: string
  menuItemPriceTextTransform?: string
  menuItemTableDataLineHeight?: string
  menuItemTableDataPadding?: string
  menuItemTableFontFamily?: string
  menuItemTablePadding?: string
  menuItemTableWidth?: string
  menuItemWidth?: string
  menuMargin?: string
  menuPadding?: string
  menuWarningColor?: string
  menuWarningFontFamily?: string
  menuWarningFontSize?: string
  menuWarningFontWeight?: string
  menuWarningLineHeight?: string
  menuWarningMargin?: string
  menuWarningMaxWidth?: string
  menuWarningPadding?: string
  menuWarningTextTransform?: string
  sideItemBackground?: string
  sideItemBorder?: string
  sideItemBorderRadius?: string
  sideItemBoxShadow?: string
  sideItemColor?: string
  sideItemDisplay?: string
  sideItemFlexFlow?: string
  sideItemFontFamily?: string
  sideItemFontSize?: string
  sideItemFontWeight?: string
  sideItemJustifyContent?: string
  sideItemLetterSpacing?: string
  sideItemLineHeight?: string
  sideItemMargin?: string
  sideItemMaxWidth?: string
  sideItemPadding?: string
  sideItemTextAlign?: string
  sideItemTextShadow?: string
  sideItemTextTransform?: string
  sideItemWidth?: string
  sides?: any
  sidesContainerDisplay?: string
  sidesContainerFlexFlow?: string
  sidesContainerJustifyContent?: string
  sidesContainerMaxWidth?: string
  sidesContainerOrderNumber?: string
  sidesContainerTitleText?: string
}